Change label show page
